+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Added Packages
+  * [Mplatform](https://pub.dev/packages/mplatform)
+  * [Awesome Notifications](https://pub.dev/packages/awesome_notifications)
+  * [Local Notifier](https://pub.dev/packages/local_notifier)
+  * [Elegant Notifications](https://pub.dev/packages/elegant_notification)
+* Provided DesktopOptions for customizing desktop notifications
+* Provided MobileOptions for customizing mobile notifications
+* Provided WebOptions for customizing web notifications
+* Init method requires a context to show the permission rational for MobileOptions